HDPI-1476 - Reorganising the fee work
