+++ src/main/java/uk/gov/hmcts/reform/pcs/payment/fee/config/FeesConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.payment.config;
+package uk.gov.hmcts.reform.pcs.payment.fee.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
